+++ components/Field/index.js
@@ -31,6 +31,7 @@ const validateIcon = props => {
 };
 
 const Container = styled.div`
+  width: 100%;
   margin-top: ${props => getFromProps(props, 'marginTop', 0)};
   margin-bottom: ${props => getFromProps(props, 'marginBottom', props.validation === false ? 12 : 32)};
   margin-left: ${props => getFromProps(props, 'marginLeft', 0)};
@@ -41,9 +42,9 @@ const StyledLabel = styled.label`
   font-weight: 400;
   font-family: ${props => props.fontFamily ? props.fontFamily : props.theme.fonts.fontPrimary};
   color: ${props => getFromProps(props, 'fontColor', props.theme.colors.neutral5)};
-  text-transform: capitalize;
 `;
 const InputContainer = styled.div`
+  background-color: white;
   width: calc(100% - 2px);
   height: 40px;
   border-width: 1px;
@@ -113,7 +114,8 @@ const Field = ({
   fontColor,
   iconColor,
   validation,
-  validationMessage
+  validationMessage,
+  placeholder
 }) => {
 
   const handleChange = event => {
@@ -125,11 +127,15 @@ const Field = ({
     return React.cloneElement(icon, {onClick: iconHasOnClick ? iconHasOnClick : onIconClick});
   };
 
+  const capitalizeFirstLetter = string => {
+    return string.charAt(0).toUpperCase() + string.slice(1);
+  };
+
   return (
     <Container marginTop={marginTop} marginRight={marginRight} marginBottom={marginBottom} marginLeft={marginLeft} validation={validation}>
-      <StyledLabel fontColor={fontColor}>{label}</StyledLabel>
+      {label && <StyledLabel fontColor={fontColor}>{capitalizeFirstLetter(label)}</StyledLabel>}
       <InputContainer radius={radius} activeColor={activeColor} validation={validation} borderColor={borderColor} iconColor={iconColor}>
-        <StyledInput type={type} value={value} onChange={handleChange} radius={radius} fontFamily={fontFamily} activeColor={activeColor} borderColor={borderColor} validation={validation} fontColor={fontColor} />
+        <StyledInput type={type} value={value} onChange={handleChange} radius={radius} fontFamily={fontFamily} activeColor={activeColor} borderColor={borderColor} validation={validation} fontColor={fontColor} placeholder={placeholder} />
         {icon && getIconFromProps()}
       </InputContainer>
       {validation === false && <ErrorMessage>{validationMessage}</ErrorMessage>}
@@ -138,6 +144,10 @@ const Field = ({
 };
 
 Field.propTypes = {
+  /**
+  *Corresponde ao placeholder
+  */
+  placeholder: PropTypes.string,
   /**
    * Corresponde a um margin-top
    */