+++ index.js
@@ -762,7 +762,20 @@ var Button = function Button(_ref) {
       size = _ref.size,
       style = _ref.style,
       variant = _ref.variant,
-      weight = _ref.weight;
+      weight = _ref.weight,
+      loading = _ref.loading,
+      enterKey = _ref.enterKey;
+  var reference = React.useRef();
+  React.useEffect(function () {
+    var handleEvent = function handleEvent(event) {
+      if (event.keyCode === 13) {
+        event.preventDefault();
+        enterKey();
+      }
+    };
+
+    reference.current.addEventListener('onEnterKey', handleEvent);
+  });
   var getClass = classnames__default['default']((_classnames = {
     'Prensa-Button-root': true
   }, _defineProperty(_classnames, "size-".concat(size, " ").concat(variant, " color-").concat(color, " radius-").concat(radius), true), _defineProperty(_classnames, 'disabled', disabled), _defineProperty(_classnames, 'fullWidth', fullWidth), _defineProperty(_classnames, 'has-leftIcon', leftIcon), _defineProperty(_classnames, 'has-rightIcon', rightIcon), _defineProperty(_classnames, "".concat(className), className), _classnames));
@@ -778,15 +791,21 @@ var Button = function Button(_ref) {
     return 'white';
   };
 
+  var getChildren = function getChildren() {
+    if (loading) return 'Carregando...';
+    return /*#__PURE__*/React__default['default'].createElement(React__default['default'].Fragment, null, leftIcon && leftIcon, children, rightIcon && rightIcon);
+  };
+
   return /*#__PURE__*/React__default['default'].createElement("button", {
     className: getClass,
     disabled: disabled,
     onClick: !disabled && onClick,
-    style: style
+    style: style,
+    ref: reference
   }, /*#__PURE__*/React__default['default'].createElement(ButtonTypography, {
     color: getFontColor(),
     weight: weight
-  }, leftIcon && leftIcon, children, rightIcon && rightIcon));
+  }, getChildren()));
 };
 
 Button.propTypes = {
@@ -806,7 +825,9 @@ Button.propTypes = {
   rightIcon: PropTypes__default['default'].oneOf([PropTypes__default['default'].object, PropTypes__default['default'].element]),
   size: PropTypes__default['default'].oneOf([1, 2, 3, 4, 5]),
   style: PropTypes__default['default'].object,
-  weight: PropTypes__default['default'].oneOf(weights)
+  weight: PropTypes__default['default'].oneOf(weights),
+  loading: PropTypes__default['default'].bool,
+  enterKey: PropTypes__default['default'].func
 };
 Button.defaultProps = {
   disabled: false,
@@ -814,7 +835,8 @@ Button.defaultProps = {
   color: 'primary-1',
   radius: 'default',
   size: 1,
-  weight: 'regular'
+  weight: 'regular',
+  loading: false
 };
 
 var pathToImage = function pathToImage(derivative, domain, policy_id, width) {
@@ -1007,7 +1029,7 @@ function _templateObject4() {
 }
 
 function _templateObject3() {
-  var data = _taggedTemplateLiteral(["\n  width: calc(100% - 2px);\n  height: 40px;\n  border-width: 1px;\n  border-style: solid;\n  border-color: ", ";\n  border-radius: ", ";\n  display: flex;\n  align-items: center;\n  svg {\n    width: 32px;\n    height: 32px;\n    margin-right: 8px;\n    fill: ", ";\n    cursor: pointer;\n  }\n  &:focus-within {\n    border-color: ", ";\n    border-width: 2px;\n    width: calc(100% - 3px);\n    height: calc(40px - 2px);\n  }\n"]);
+  var data = _taggedTemplateLiteral(["\n  background-color: white;\n  width: calc(100% - 2px);\n  height: 40px;\n  border-width: 1px;\n  border-style: solid;\n  border-color: ", ";\n  border-radius: ", ";\n  display: flex;\n  align-items: center;\n  svg {\n    width: 32px;\n    height: 32px;\n    margin-right: 8px;\n    fill: ", ";\n    cursor: pointer;\n  }\n  &:focus-within {\n    border-color: ", ";\n    border-width: 2px;\n    width: calc(100% - 3px);\n    height: calc(40px - 2px);\n  }\n"]);
 
   _templateObject3 = function _templateObject3() {
     return data;
@@ -1017,7 +1039,7 @@ function _templateObject3() {
 }
 
 function _templateObject2() {
-  var data = _taggedTemplateLiteral(["\n  font-size: 14px;\n  font-weight: 400;\n  font-family: ", ";\n  color: ", ";\n  text-transform: capitalize;\n"]);
+  var data = _taggedTemplateLiteral(["\n  font-size: 14px;\n  font-weight: 400;\n  font-family: ", ";\n  color: ", ";\n"]);
 
   _templateObject2 = function _templateObject2() {
     return data;
@@ -1027,7 +1049,7 @@ function _templateObject2() {
 }
 
 function _templateObject() {
-  var data = _taggedTemplateLiteral(["\n  margin-top: ", ";\n  margin-bottom: ", ";\n  margin-left: ", ";\n  margin-right: ", ";\n"]);
+  var data = _taggedTemplateLiteral(["\n  width: 100%;\n  margin-top: ", ";\n  margin-bottom: ", ";\n  margin-left: ", ";\n  margin-right: ", ";\n"]);
 
   _templateObject = function _templateObject() {
     return data;
@@ -1112,7 +1134,8 @@ var Field = function Field(_ref) {
       fontColor = _ref.fontColor,
       iconColor = _ref.iconColor,
       validation = _ref.validation,
-      validationMessage = _ref.validationMessage;
+      validationMessage = _ref.validationMessage,
+      placeholder = _ref.placeholder;
 
   var handleChange = function handleChange(event) {
     onChange(event.target.value);
@@ -1125,6 +1148,10 @@ var Field = function Field(_ref) {
     });
   };
 
+  var capitalizeFirstLetter = function capitalizeFirstLetter(string) {
+    return string.charAt(0).toUpperCase() + string.slice(1);
+  };
+
   return /*#__PURE__*/React__default['default'].createElement(Container, {
     marginTop: marginTop,
     marginRight: marginRight,
@@ -1133,7 +1160,7 @@ var Field = function Field(_ref) {
     validation: validation
   }, /*#__PURE__*/React__default['default'].createElement(StyledLabel, {
     fontColor: fontColor
-  }, label), /*#__PURE__*/React__default['default'].createElement(InputContainer, {
+  }, capitalizeFirstLetter(label)), /*#__PURE__*/React__default['default'].createElement(InputContainer, {
     radius: radius,
     activeColor: activeColor,
     validation: validation,
@@ -1148,11 +1175,17 @@ var Field = function Field(_ref) {
     activeColor: activeColor,
     borderColor: borderColor,
     validation: validation,
-    fontColor: fontColor
+    fontColor: fontColor,
+    placeholder: placeholder
   }), icon && getIconFromProps()), validation === false && /*#__PURE__*/React__default['default'].createElement(ErrorMessage, null, validationMessage));
 };
 
 Field.propTypes = {
+  /**
+  *Corresponde ao placeholder
+  */
+  placeholder: PropTypes__default['default'].string,
+
   /**
    * Corresponde a um margin-top
    */